moved script to script foleder for better readabilty
